agents/vibe-coder/requirements.txt
fastapi==0.110
uvicorn==0.29
nats-py==2.6
PyGithub==2.3
